Seperate logic into index.js and server module
